Added Preset Prozesses and Task 2

+++ Scheduling/EarliestDeadlineFirst.cpp
@@ -5,7 +5,7 @@ void EarliestDeadlineFirst::Schedule() {
     sizeAll = m_Prozesses.size();
 
     while (prozesseDone.size() != sizeAll) {
-        for (int i = 0; i < m_Prozesses.size(); ++i) {
+        for (size_t i = 0; i < m_Prozesses.size(); ++i) {
             if (m_Prozesses[i].m_timeArrival <= timeNow && !m_Prozesses[i].dead) {
                 m_Prozesses[i].dead=true;
                 prozesseOpen.push_back(m_Prozesses[i]);
@@ -23,7 +23,7 @@ void EarliestDeadlineFirst::Schedule() {
             timeNow++;
         }
     }
-    for (int i = 0; i < prozesseDone.size(); ++i) {
+    for (size_t i = 0; i < prozesseDone.size(); ++i) {
         prozesseDone[i].m_timeTookToCalculate = prozesseDone[i].m_doneTime - prozesseDone[i].m_timeArrival;
     }
     for (size_t i = 0; i < prozesseDone.size(); i++) {
